--- optimizations/preallocate/tiny_eval_output.txt
@@ -1,36 +1,0 @@
-nparams = 1985146305
-len(preds) = 4, elapsed = 10.858907461166382, elapsed / len(preds) = 2.7147268652915955
-len(preds) = 4, elapsed = 0.040189504623413086, elapsed / len(preds) = 0.010047376155853271
-len(preds) = 4, elapsed = 0.0398867130279541, elapsed / len(preds) = 0.009971678256988525
-len(preds) = 4, elapsed = 0.03987765312194824, elapsed / len(preds) = 0.00996941328048706
-len(preds) = 4, elapsed = 0.039879560470581055, elapsed / len(preds) = 0.009969890117645264
-len(preds) = 4, elapsed = 0.03997611999511719, elapsed / len(preds) = 0.009994029998779297
-len(preds) = 4, elapsed = 0.03976130485534668, elapsed / len(preds) = 0.00994032621383667
-len(preds) = 4, elapsed = 0.03981280326843262, elapsed / len(preds) = 0.009953200817108154
-len(preds) = 4, elapsed = 0.03980517387390137, elapsed / len(preds) = 0.009951293468475342
-len(preds) = 4, elapsed = 0.039873361587524414, elapsed / len(preds) = 0.009968340396881104
-len(preds) = 4, elapsed = 0.03991055488586426, elapsed / len(preds) = 0.009977638721466064
-len(preds) = 4, elapsed = 0.03977489471435547, elapsed / len(preds) = 0.009943723678588867
-len(preds) = 4, elapsed = 0.03983926773071289, elapsed / len(preds) = 0.009959816932678223
-len(preds) = 4, elapsed = 0.03980517387390137, elapsed / len(preds) = 0.009951293468475342
-len(preds) = 4, elapsed = 0.03997230529785156, elapsed / len(preds) = 0.00999307632446289
-len(preds) = 4, elapsed = 0.03981900215148926, elapsed / len(preds) = 0.009954750537872314
-len(preds) = 4, elapsed = 0.039792776107788086, elapsed / len(preds) = 0.009948194026947021
-len(preds) = 4, elapsed = 0.03979349136352539, elapsed / len(preds) = 0.009948372840881348
-len(preds) = 4, elapsed = 0.03979682922363281, elapsed / len(preds) = 0.009949207305908203
-len(preds) = 4, elapsed = 0.0398101806640625, elapsed / len(preds) = 0.009952545166015625
-len(preds) = 4, elapsed = 0.03981637954711914, elapsed / len(preds) = 0.009954094886779785
-len(preds) = 4, elapsed = 0.039755821228027344, elapsed / len(preds) = 0.009938955307006836
-len(preds) = 4, elapsed = 0.03976297378540039, elapsed / len(preds) = 0.009940743446350098
-len(preds) = 4, elapsed = 0.039775848388671875, elapsed / len(preds) = 0.009943962097167969
-len(preds) = 4, elapsed = 0.0397794246673584, elapsed / len(preds) = 0.0099448561668396
-
-Total requests: 100
-Total responses: 100
-Response rate: 100.00%
-Average latency: 472.7 ms
-Tower Accuracies:
-XLSTM: max_error = 0.0000, mean_error = 0.0000
-Mamba2: max_error = 0.0000, mean_error = 0.0000
-RetNet: max_error = 0.0000, mean_error = 0.0000
-Hawk: max_error = 0.0000, mean_error = 0.0000
